arm64: dts: qcom: c630: keep both touchpad devices enabled
Indicated by AML code in ACPI table, the touchpad in-use could be found
on two possible slave addresses on &i2c3, i.e. hid@15 and hid@2c.  And
which one is in-use can be determined by reading another address on the
I2C bus.  Unfortunately, for DT boot, there is currently no support in
firmware to make this check and patch DT accordingly.  This results in
a non-functional touchpad on those C630 devices with hid@2c.

As i2c-hid driver will stop probing the device if there is nothing on
the slave address, we can actually keep both devices enabled in DT, and
i2c-hid driver will only probe the existing one.  The only problem is
that we cannot set up pinctrl in there, as two devices with the same
pinctrl will cause pin conflict that makes the second device probe fail.
Considering the pin setup has been done by firmware for the device, it's
not a problem in practice.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>