HACK: ACPI: tables: trick of oem_revision check
We need to use initrd upgrade mechanism to fix GpioInt number of TSC2
(trackpad) in Lenovo Flex 5G DSDT table.  It works, however, as we run
a single initrd in installer for all laptops, including Lenovo Yoga
C630, the DSDT upgrade will undesirably apply on Yoga C630 as well.
That's because the DSDT table of Flex 5G and Yoga C630 share exactly
same signature, oem_id and oem_table_id, so the upgrade check in kernel
pass for both laptops.

This is a hack to trick the check of oem_revision filed which is
different between Flex 5G and Yoga C630 DSDT table, so that the upgrade
only happens on Flex 5G with the same oem_revision.

The real fix should be DSDT table itself.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>